[Sua Peca - Web] - #333: Ao calcular um frete com as dimensoes 1000,1000,1000 - 5kg e preco 123,45.

+++ VIKomet/SDK/Entities/WebStore/Freight/FreightTable.cs
@@ -33,7 +33,7 @@ namespace VIKomet.SDK.Entities.VIFrete
         public DateTime DateTo { get; set; }
 
         [DataMember(Name = "CubedWeightFactor")]
-        public decimal CubedWeightFactor { get; set; }
+        public double CubedWeightFactor { get; set; }
 
         [DataMember(Name = "CreateDate")]
         public DateTime CreateDate { get; set; }
